Add a constant folding and constant propagation pass
The main goal for this pass is to eliminate some type conversion
functions that would be introduced for lack of sufficient syntax
to express certain types (like locations).
But it's a simple optimization and it's good to have anyway.

Besides, after all the prep work it's not that much code to
implement.