Make type conversion calls work in input arguments
This was the purpose of the whole patch series, and requires
moving the check for "var ref or constant" after constant folding
converts function calls to constants.
Tests check that the behavior is still the expected one.

Also add support for codegenning locations, pictures and resources,
which Sabrina will need.