Reimplement compiler on top of leaner AST and grammar
Cut down the grammar to match exactly what ThingTalk wants to do
(plus multiple queries, plus multiple actions), and simplify the
AST nodes accordingly.
Eliminate compute modules and stored variables, which we don't need
for now, and will reappear in a different form if we ever need them.
Cut down on output computation, which is not usable anyway.