Compile ThingTalk to JavaScript code instead of combining closures
Generate actual JS code to be passed to eval. This is a little
slower at compile time, but it will be faster at runtime for long-running
rules (because the JIT has a fighting chance at doing something),
and it will make it easier to implement richer and more complex
filters.