Compiler: fix filters on input parameters
Filters on input parameters are somewhat of a weird case, because
input parameters are specified and so filters are usually redundant.
Sometimes though they are not, or eliminating the redundancy is
difficult, especially when the filter is added by a policy, and
therefore they are supported by the language. Therefore they
should be supported by the compiler.