SchemaRetriever: consolidate and simplify a lot code dealing with schemas
Change the API to always return Ast.ClassDef and Ast.FunctionDef
directly from SchemaRetriever.

This is a pretty big change for users of SchemaRetriever, but luckily
they all live in almond-cloud, which is changing at the same time.

We will need to coordinate the change across thingpedia-client,
almond-cloud and here.