Describe: change how "when => get => " programs are expressed
The form "$action when $stream then $query" does not work, and the form
"$query then $action when $stream" is hard to implement and also weird.

Instead, do the plain "when $stream, $query and then $action".

To ensure that the whole sentence in Almond reads, including the
part "Ok, so you want me to", we add the prefix "do the following:",
so overall it reads as:

"Ok, so you want me to do the following: when $when, $query and then $action".

Fixes #12