Add versioning info to the API
Use ThingTalk.version to learn the version of the library

This commit duplicates the info from package.json to index.js,
as otherwise we have to load the whole package.json (using require
to be browserify compatible), which is a waste of memory to keep
for just 3 bytes.
A test ensures that the two places don't go out of sync.