classdefast: migrate any entity type to input[type=text]
If we don't know about a certain entity type, we should just treat
it as text input.

In practice, the concrete value matters only in obscure compatibility
cases, because modern code will use the factory coming out of Thingpedia
instead, but it is important that we don't crash when the user
types an arbitrary Entity type in their mixin.