Add a python backend to thingtalk-generate-parser
Invoke with `thingtalk-generate-parser -l python`

This will generate an SLR parser that can be used with genie-parser
(luinet-parser, almond-nnparser, or whatever it's called these
days).

Note that if you run this on the ThingTalk nn grammar, the result
is a significantly looser grammar than the one currently used
by the parser, because it does not depend on Thingpedia at all,
and it does not check that each constant has the type that the
parameter expects.

This is for ease of writing the grammar by hand, and it has a negligible
effect if you use linear decoding (the effect with topdown/bottomup decoding
could be larger). If we care, we can in the future write a tighter
grammar with type information.