Change the ABI (ThingTalk-to-Thingengine) to support multi-command
Instead of returning multiple independent compiled statements,
now AppCompiler returns a single compiled object containing a
"command" function (to be executed now and once) and an array
of "rule" functions (to be executed every time the engine is restarted).

In particular, this means there is a single shared namespace of
state IDs across the whole program, which matches what
thingengine-core was inadvertendly doing already anyway.