builtin/equality: handle comparisons between Dates and Strings
JS Date objects don't survive JSON serialization, they are converted
to string. Yet, we rely on JSON serialization to store previous
polling results for monitor(), and we use "equality()" to check
if a notification is duplicate. We thus need to handle comparisons
between strings and dates at runtime.

This is OK because ThingTalk is strongly typed so the two values
must have the same TT type, even if their in-memory representation
is different.