ToNN: add the ability to allocate the entities sequentially
When passing the context to the neural network, we want a representation
that does not depend on the input sentence, in particular in which order
entities where represented in the sentence. Instead, we want
a normalized, sequential order.

This adds an option to toNN that uses a different entity assignment
strategy, instead of fetching from a precomputed bag.