Introduce unresolved locations
Currently, locations are preprocessed and resolved (mapped to a concrete
point on Earth) by the tokenizer. This is not great, because it
relies on CoreNLP's NER for locations (which has poor quality and is
inefficient to run), and because ambiguous locations cannot be resolved
by the user.

Instead, we can treat locations in manner similar to entities:
copy the tokens from the input into the program, and defer entity
linking to the dialog agent during slot filling.
Among other things, this will also enable personalized location
databases.