More refined canonical annotations for parameters
By analyzing the turking and developer data for question answering, we
realize more refined annotations are needed in order to make realistic
compositions.
We introduce 6 different categories, which can be found in:
https://wiki.almond.stanford.edu/LUInet/QuestionAnswering

Now canonical takes either a string or an object.
describe.js currently only takes the noun-phrase canonical for property (NPP)
when an object is given. Future improvement is needed to fully leverage
all the information and create more natural confirmation.