Add a new API to iterate slots
The existing slot iteration API is hairy, with 4 values that are
returned with a plain array. Furthermore, the slot is returned
directly and the calling code is expected to read/set `.value`, which
is problematic when the value is not in `.value` (eg. an index
of an array, or the time of a timer).

To solve both of these problems, introduce a new slot iteration API,
which yield objects with .get() and .set() methods. These objects
also know their own .type (and soon their own .prompt), which removes
some duplicated code in genie and almond-dialog-agent.

For compatibility, the new API is called `iterateSlots2`, and the
existing `iterateSlots` is not affected. This might change in the future,
after all callers have been updated.