Add `field` to aggregate filter to handle non-count operators
The original design was to do things like `avg(ratings)` if `ratings` is an array of scalar values;
and `avg(ratings.value)` if `ratings` is an array of compound values.

This will be problematic since we flatten compound values when creating the class - we will need
to check if `ratings.value` is flattened parameter or an array of compound. It's even worse when
multiple dots are in the name, e.g., `aggregateRating.reviews.ratingvValue`.

Thus, we follow the similar syntax as in aggregate table: `avg(value of ratings)`