Add device attributes to device selectors
Attributes are key-value pairs that allow the user to select a device by
name, location, or other characteristic.

Currently, the only allowed attribute is name, of String type.

The AST is more flexible though to support parameter passing (needed
to make primitive templates with device names). $? is not supported
because it's not very useful - Almond natively asks for which
device anyway.