Add a visitor pattern and class to traverse AST
To traverse AST without knowledge of the internal structure,
introduce a new Ast.NodeVisitor class. Subclassing this class
allows to handle specific types of nodes easily, as well as
traverse all nodes in the tree. It can subsubsume iterateSlots
and iteratePrimitive, and it can be a way to implement iterateSlots2.