Make visiting ASTs synchronous
I originally (in almondscript) implemented async visiting to support
typechecking, which must call into SchemaRetriever's async methods.
But sync visiting is useful to reimplement some iterate* APIs
in terms of visitors, and it is also useful in contexts where
async is not available (eg Genie). It is also very slightly faster.