Merge Ast.Value and Ast.ScalarExpression
Ast.Value already included some values that are not constants
(variables, context references, array literals), but it does not
include others (math), which are exposed by ScalarExpression.
This is confusing, limiting, and leads to clunky syntax, especially
around "list expression"

Given that compute expressions are fairly new and we haven't used
them too much, we have a chance to clean this up.
Gone are ScalarExpression and ListExpression. Instead we have:

- Value.Computation: for math, distance and aggregation ops
- Value.ArrayField: to extract a field of each element of an array of records
- Value.Filter: to filter an array

All three combined yield the same power as ScalarExpression + ListExpression,
with mostly the same syntax.

For reason of grammar ambiguity, Filter is expressed as "val filter { foo }".
We also change window/timeseries/sequence/history to use parenthesis
around their arguments. This is a breaking change, but those are never fully
implemented experimental features.
Compute filters and compute expression change in syntax too, slightly.