Introduce "new" syntax for constructing constants
Originally proposed for almondscript, it replaces the "make"
pseudo-factory functions: instead of `makeDate(...)` you can
now write `new Date(...)` etc.
(The old syntax is still supported to help developers who learned it)

Also, declare all type names to be keywords. This will help with
the long-term goal of uniforming NN syntax and surface syntax, because
NN syntax needs to be unambiguous wrt keywords, as it uses a deterministic
parser.