Nerf typechecking of natural language annotations
These annotations are dependent on the specific Genie templates
and Genie version used, and the whole point of using annotations
is that we don't need to update the version of ThingTalk to
change how annotations work. So the typecheck here is misplaced.

Thingpedia should check these before they go public (like we check
other annotations), perhaps using some API in Genie.