dialogue: extend syntax of dialogue results to track the number of results
We don't want to expose the full result list to the neural network,
but we want to expose the number, so the neural network can surface
it in the resulting sentence, like `I found 55 results, here are the
top two:`.
Also, add a bit for `more`, indicating the count is a lower bound
on the true number of results, so we don't have to fully evaluate
a statement, and we can be lazy.