Fix typechecking enum arrays
The array [enum(a), enum(b)] is composed of:
- enum(a) with type Enum(a, *)   ("enum of a and other values")
- enum(b) with type Enum(b, *)   ("enum of b and other values")

The two types are not equal, so this was previously reported as
inconsistent type for the array. Instead, the array is valid and
should have type Enum(a, b, *) ("enum of a, b and other values").

We started having this problem because we started converting
"x == enum(a) || x == enum(b)" to "in_array(x, [enum(a), enum(b)])").