Add DatePiece class
DatePiece is a way to construct Date values when only one of [year,
month, day] is specified. The big-end non-supplied values get set
to the present and the little-end non-supplied values get set to the
smallest possible. For example: DatePiece('month', 3) gets compiled to
the Date corresponding to the first of April in the present year
(January is month 0). This is useful for disambiguating a reference
such as "the 11th" in the context of dates.