Split NNSyntax for DatePiece into two rules
Now there is one rule for time === -1, and one rule for time !== -1 (or
more specifically, one rule expecting a number, and one expeciting a
date). With this, we remove the rule time:-1 => Ast.Value.Number(-1) in
constant_Time.

Also fix NNSyntax for -1 in yr/mo/dt (was '-1', now is '-', '1', as is
the general pattern in NNSyntax).

Also fix some linter errors (missing curly braces)