Remove DatePiece syntax (TT), fix conversion to NN
The `new DatePiece(...)` syntax should have been removed from ThingTalk
as per the previous two commits (it is done now).

Also, in converting years to NN syntax, the previous code was
searching for the last two digits of any year greater than 1950, but in
reality we wanted to do that only for years between 1950 and 2050. This
commit fixes that.