Add support for partial Date syntax (#232)
Now in ThingTalk one may create Date objects without explicitly specifying
all the parameters. Missing parameters are assumed to be either the present
(if to the left of the provided parameters), or the smallest possible (if to the right
of the provided parameters). Additionally, two-digit years are assumed to be
abbreviations for years between 1950 and 2050 (excluded). For example:
new Date(95, 2, ) refers to the first of February in 1995. 
new Date(, , 28, 7, 30, 0) refers to 7:30 on the 28th of the current month and year.

The NN syntax for such dates omits the missing parameters, and represents time as a
TIME entity, to be found in context.
The AST Date value represents this by use of the DatePiece class (introduced here).