Fill in the new syntax
To reduce redundancy, the new syntax does not use "now" and "notify".
Hence, we cannot recognize if something is a stream, table, or action
statically at parse time.

To support the new syntax, we introduce new AST datastructures based
on a uniform Expression class. The type of Expression (stream, table,
action) is assigned during typechecking through ExpressionSignature.

This commit adds the new AST structures, and the syntax for simple
statements, expressions, values, and types.

Typechecking and the result of the language will be added later.