Make the new syntax the default
Remove prettyprinting in the old syntax, switch Grammar.parse()
to take new syntax by default, and add APIs to parse and
serialize to and from the new syntax and the tokenized version
of the new syntax (formely known as NN syntax).

This is a large commit because it ports over all the tests to
the new syntax, and then fixes everything to make those darn
tests pass.