Introduce * projections
A projection of the form [*, foo] where foo is a computations
adds the computation value without removing any output parameter.

This feature is necessary to translate old compute tables correctly,
which in turn is necessary to deal with existing datasets in Genie,
in particular around sort of computations.

Also, expand the normalizations around projections and computes,
and add tests.