Move source filters on policy to use a dedicated $source variable
This change arises from a bug when replacing parameters in programs
that apply a filter on a join: the existing FilterSlot assumed
that any filter on something without a primitive would be a source
filter. A join does not have a primitive, so this breaks.

To fix that, we change the "source" variable to "$source".
$ variables are Ast.EventValue instead of Ast.VarRefValue, so
source filters need to be ComputeBooleanExpression instead of
AtomBooleanExpression. The change, and the associated compatibility
hacks to preserve existing legacy syntax that uses "source", then
cascades everywhere.