Add back the ability to serialize ThingTalk to legacy syntax
almond-cloud needs to convert new thingtalk to old thingtalk to
serve the Thingpedia API to legacy clients. Otherwise we are
forced to update all clients at once which is going to be
problematic.

In the process, fix a number of issues in the conversion between
new ASTs and legacy ASTs, and some prettyprinting issues with
nested compound types.