Optimize parsing ThingTalk code (#287)
Avoid setting .length, because that calls into a runtime function
instead of being compiled inline, and transitions the arrays to
a sparse representation (https://v8.dev/blog/elements-kinds)

Avoid calling the String() constructor because that calls into
the runtime and does a lot of checks to eventually call into
toString(). We can avoid all of that with a single type check.

Overall, this commit results in a 25% speed up in dialog-to-contextual,
which is one of the bottlenecks of synthesis for dialogues.