Make $context values non-compilable, and support nested keys (#295)
$context values are placeholders for future values. Making
them non-compilable means I can (need to) replace them in
the dialogue agent before execution.

This commit makes $context.result suitable to represent
non-default parameter passing across multiple dialogue turns.

Example: "search a website then post the title on twitter"
can be split into:
```
@com.bing.web_search();
@com.twitter.post(status=$context.result.title : String);
```
to be stored in the dialogue state.

The semantics of this are that the dialogue agent executes
the first command, and presents the result to the user.
The dialogue agent remembers that the user wanted to post the title
on Twitter, and this is preserved across slot-filling of the
query and search refinements, so when the agent offers
"do you want to post it on twitter", or when the user says
"ok now post it on twitter", the right parameter passing occurs.

I thought about using `$result` instead of `$context.result` but
`$result` already means "a string representation of the current
result" (the old `$event` in ThingTalk 1.0) and overloading the
meaning was awkward. Plus we need type annotations to typecheck
the individual dialogue turns without looking at the context.

Note that the new semantics are also more correct for `$context.selection`
(the selection on the screen), because that one also should be
resolved once and for all when the program is started, and not
delayed until the action is invoked, potentially much later.