Fix inadvertently merging location entities
Two locations that are less than 0.1 degree apart in either coordinate
can be quite far (for latitude, 0.1 degrees is about 7 miles).
Instead, we should compare a lot more strictly, so we only catch
floating point rounding errors (due to roundtripping from JSON/strings
in various places) and not locations that are actually distinct.