SchemaRetreiver: typecheck dataset examples individually
In certain cases, the dataset can fail to typecheck as a whole
because there are examples with newer functions or wrong parameters.
This can occur if a device has been updated recently (in the last
3 hours), or can occur for a builtin device if Genie or the platform
layers have not been updated yet.
In that case, we should ignore and log the error, and let as many
examples through as we can.

Fixes https://github.com/stanford-oval/thingpedia-common-devices/issues/286
and will probably help with https://github.com/stanford-oval/almond-gnome/issues/40