Stop using legacy AST nodes in the compiler
Refactor the compiler to convert Ast.Expression to TableOp/StreamOp
without going through Ast.Table/Ast.Stream/Ast.Action, which are
deprecated.

The bulk of the compiler, which deals with TableOp/StreamOp,
is unchanged. There are a couple small changes in ReduceOp because
SortExpression takes an arbitrary sort value rather than a single
field.

This fixes the compilation of sort expressions (which should not
return the sort key as additional field), fixed the query hints
with computed output parameters, and most importantly fixes
passing input parameters to user-declared functions.