Support all entities untokenized in the sentence (#357)
If an entity is not extracted by the tokenizer, try to look it up
with its tokens exactly in the sentence. Previously we did so
for hashtags and usernames, now we add the missing entities:
path names, urls, email address, phone numbers.

This fix is particularly important for path names because strings
can no longer be used as path names, and path names are typically
not identifiable by preprocessing.