Remove special syntax for join condition, use filter instead
As we are keeping the `first.`/`second.` prefix, there is no need to use
a special syntax for join condition. It also simplifies normalization.

One missing part is that we should probably always push down filters
regarding just one table in the join from outside the join to inside the
join table.
But it should be fine without it for now as at the moment, we do not plan
to generate examples with any other operations on top of join.