allow filtering directly on compound type
let's say `p` is a property of compound type, `p == 3` it's equivalent to
`p.value == 3`. The type of `value` field is used to typecheck the
filter.  If there is no `value` field available, it does not typecheck.

similarly, for projection, if there is a `value` field for a compound
type, then we only project the value; otherwise, we project the whole
object.